requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.0
SQLAlchemy==1.4.27
psycopg2==2.9.1
requests==2.26.0
openai==0.11.3
Werkzeug==2.0.1